travel designer/requirements.txt
openai
python-dotenv
chainlit 